--- src/main/java/com/hiddenmuseumdemo/service/DemoService.java
@@ -1,5 +1,0 @@
-package com.hiddenmuseumdemo.service;
-
-public interface DemoService {
-    String getDemoInfo(String medium,String objectType,String artist);
-}
