+++ src/main/java/uk/ac/bristol/hiddenmuseum/controller/SearchController.java
@@ -1,12 +1,10 @@
-package com.hiddenmuseumdemo.controller;
+package uk.ac.bristol.hiddenmuseum.controller;
 
-import com.hiddenmuseumdemo.service.DemoService;
+import uk.ac.bristol.hiddenmuseum.service.DemoService;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
