+++ src/main/java/uk/ac/bristol/hiddenmuseum/service/DemoService.java
@@ -1,0 +1,5 @@
+package uk.ac.bristol.hiddenmuseum.service;
+
+public interface DemoService {
+    String getDemoInfo(String medium,String objectType,String artist);
+}
