+++ src/test/java/uk/ac/bristol/hiddenmuseum/SkeletonApplicationTests.java
@@ -1,4 +1,4 @@
-package com.hiddenmuseumdemo;
+package uk.ac.bristol.hiddenmuseum;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
